+++ src/style.rs
@@ -1,8 +1,4 @@
-use crate::utils::*;
-use egui::{
-    ecolor::*, emath::*, epaint, Align2, CursorIcon, Id, Margin, Response, Rounding, Sense, Stroke,
-    TextStyle, Ui, WidgetText,
-};
+use egui::{ecolor::*, Margin, Rounding, Stroke};
 
 /// Left or right alignment for tab add button.
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
@@ -198,6 +194,13 @@ impl Default for TabsStyle {
 }
 
 impl Style {
+    pub(crate) const TAB_ADD_BUTTON_SIZE: f32 = 24.0;
+    pub(crate) const TAB_ADD_PLUS_SIZE: f32 = 12.0;
+    pub(crate) const TAB_CLOSE_BUTTON_SIZE: f32 = 24.0;
+    pub(crate) const TAB_CLOSE_X_SIZE: f32 = 9.0;
+}
+
+impl Style {
     /// Derives relevant fields from `egui::Style` and sets the remaining fields to their default values.
     ///
     /// Fields overwritten by [`egui::Style`] are:
@@ -259,285 +262,6 @@ impl Style {
             },
             ..Self::default()
         }
-    }
-
-    pub(crate) fn hsplit(
-        &self,
-        ui: &mut Ui,
-        fraction: &mut f32,
-        rect: Rect,
-    ) -> (Response, Rect, Rect, Rect) {
-        let pixels_per_point = ui.ctx().pixels_per_point();
-
-        let mut separator = rect;
-
-        let midpoint = rect.min.x + rect.width() * *fraction;
-        separator.min.x = midpoint - self.separator.width * 0.5;
-        separator.max.x = midpoint + self.separator.width * 0.5;
-
-        let response = ui
-            .allocate_rect(separator, Sense::click_and_drag())
-            .on_hover_and_drag_cursor(CursorIcon::ResizeHorizontal);
-
-        if let Some(pos) = response.interact_pointer_pos() {
-            let x = pos.x;
-            let delta = response.drag_delta().x;
-
-            if (delta > 0. && x > midpoint && x < rect.max.x)
-                || (delta < 0. && x < midpoint && x > rect.min.x)
-            {
-                let range = rect.max.x - rect.min.x;
-                let min = (self.separator.extra / range).min(1.0);
-                let max = 1.0 - min;
-                let (min, max) = (min.min(max), max.max(min));
-                *fraction = (*fraction + delta / range).clamp(min, max);
-            }
-        }
-
-        let midpoint = rect.min.x + rect.width() * *fraction;
-        separator.min.x = map_to_pixel(
-            midpoint - self.separator.width * 0.5,
-            pixels_per_point,
-            f32::round,
-        );
-        separator.max.x = map_to_pixel(
-            midpoint + self.separator.width * 0.5,
-            pixels_per_point,
-            f32::round,
-        );
-
-        (
-            response,
-            rect.intersect(Rect::everything_right_of(separator.max.x)),
-            separator,
-            rect.intersect(Rect::everything_left_of(separator.min.x)),
-        )
-    }
-
-    pub(crate) fn vsplit(
-        &self,
-        ui: &mut Ui,
-        fraction: &mut f32,
-        rect: Rect,
-    ) -> (Response, Rect, Rect, Rect) {
-        let pixels_per_point = ui.ctx().pixels_per_point();
-
-        let mut separator = rect;
-
-        let midpoint = rect.min.y + rect.height() * *fraction;
-        separator.min.y = midpoint - self.separator.width * 0.5;
-        separator.max.y = midpoint + self.separator.width * 0.5;
-
-        let response = ui
-            .allocate_rect(separator, Sense::click_and_drag())
-            .on_hover_and_drag_cursor(CursorIcon::ResizeVertical);
-
-        if let Some(pos) = response.interact_pointer_pos() {
-            let y = pos.y;
-            let delta = response.drag_delta().y;
-
-            if (delta > 0. && y > midpoint && y < rect.max.y)
-                || (delta < 0. && y < midpoint && y > rect.min.y)
-            {
-                let delta = response.drag_delta().y;
-                let range = rect.max.y - rect.min.y;
-                let min = (self.separator.extra / range).min(1.0);
-                let max = 1.0 - min;
-                let (min, max) = (min.min(max), max.max(min));
-                *fraction = (*fraction + delta / range).clamp(min, max);
-            }
-        }
-
-        let midpoint = rect.min.y + rect.height() * *fraction;
-        separator.min.y = map_to_pixel(
-            midpoint - self.separator.width * 0.5,
-            pixels_per_point,
-            f32::round,
-        );
-        separator.max.y = map_to_pixel(
-            midpoint + self.separator.width * 0.5,
-            pixels_per_point,
-            f32::round,
-        );
-
-        (
-            response,
-            rect.intersect(Rect::everything_above(separator.min.y)),
-            separator,
-            rect.intersect(Rect::everything_below(separator.max.y)),
-        )
-    }
-
-    pub(crate) const TAB_ADD_BUTTON_SIZE: f32 = 24.0;
-    pub(crate) const TAB_ADD_PLUS_SIZE: f32 = 12.0;
-    pub(crate) const TAB_CLOSE_BUTTON_SIZE: f32 = 24.0;
-    pub(crate) const TAB_CLOSE_X_SIZE: f32 = 9.0;
-
-    pub(crate) fn tab_plus(&self, ui: &mut Ui) -> Response {
-        let (rect, mut response) = ui.allocate_exact_size(ui.available_size(), Sense::click());
-
-        response = response.on_hover_cursor(CursorIcon::PointingHand);
-
-        let color = if response.hovered() {
-            ui.painter()
-                .rect_filled(rect, Rounding::none(), self.buttons.add_tab_bg_fill);
-            self.buttons.add_tab_active_color
-        } else {
-            self.buttons.add_tab_color
-        };
-
-        let mut plus_rect = rect;
-
-        rect_set_size_centered(&mut plus_rect, Vec2::splat(Self::TAB_ADD_PLUS_SIZE));
-
-        ui.painter().line_segment(
-            [plus_rect.center_top(), plus_rect.center_bottom()],
-            Stroke::new(1.0, color),
-        );
-        ui.painter().line_segment(
-            [plus_rect.right_center(), plus_rect.left_center()],
-            Stroke::new(1.0, color),
-        );
-
-        response
-    }
-
-    /// * `active` means "the tab that is opened in the parent panel".
-    /// * `focused` means "the tab that was last interacted with".
-    ///
-    /// Returns the main button response plus the response of the close button, if any.
-    #[allow(clippy::too_many_arguments)]
-    pub(crate) fn tab_title(
-        &self,
-        ui: &mut Ui,
-        label: WidgetText,
-        focused: bool,
-        active: bool,
-        is_being_dragged: bool,
-        id: Id,
-        expanded_width: f32,
-        show_close: bool,
-    ) -> (Response, Option<Response>) {
-        let rounding = self.tabs.rounding;
-
-        let galley = label.into_galley(ui, None, f32::INFINITY, TextStyle::Button);
-
-        let x_spacing = 8.0;
-
-        let text_width = galley.size().x + 2.0 * x_spacing;
-        let close_button_size = if show_close {
-            Self::TAB_CLOSE_BUTTON_SIZE.min(self.tab_bar.height)
-        } else {
-            0.0
-        };
-        let minimum_width = text_width + close_button_size;
-
-        // Compute total width of the tab bar
-        let tab_width = if self.tabs.fill_tab_bar {
-            expanded_width
-        } else {
-            minimum_width
-        }
-        .at_least(minimum_width);
-
-        let (rect, mut response) =
-            ui.allocate_exact_size(vec2(tab_width, ui.available_height()), Sense::hover());
-        if !ui.memory(|mem| mem.is_anything_being_dragged()) {
-            response = response.on_hover_cursor(CursorIcon::Grab);
-        }
-
-        if active {
-            if is_being_dragged {
-                ui.painter()
-                    .rect_stroke(rect, rounding, Stroke::new(1.0, self.tabs.outline_color));
-            } else {
-                let stroke = Stroke::new(1.0, self.tabs.outline_color);
-                ui.painter().rect(rect, rounding, self.tabs.bg_fill, stroke);
-
-                // Make the tab name area connect with the tab ui area:
-                ui.painter().hline(
-                    rect.x_range(),
-                    rect.bottom(),
-                    Stroke::new(2.0, self.tabs.bg_fill),
-                );
-            }
-        }
-
-        let mut text_rect = rect;
-        text_rect.set_width(tab_width - close_button_size);
-
-        let text_pos = if self.tabs.fill_tab_bar {
-            let mut pos =
-                Align2::CENTER_CENTER.pos_in_rect(&text_rect.shrink2(vec2(x_spacing, 0.0)));
-            pos -= galley.size() / 2.0;
-            pos
-        } else {
-            let mut pos = Align2::LEFT_CENTER.pos_in_rect(&text_rect.shrink2(vec2(x_spacing, 0.0)));
-            pos.y -= galley.size().y / 2.0;
-            pos
-        };
-
-        let override_text_color = if galley.galley_has_color {
-            None // respect the color the user has chosen
-        } else {
-            Some(match (active, focused) {
-                (false, false) => self.tabs.text_color_unfocused,
-                (false, true) => self.tabs.text_color_focused,
-                (true, false) => self.tabs.text_color_active_unfocused,
-                (true, true) => self.tabs.text_color_active_focused,
-            })
-        };
-
-        ui.painter().add(epaint::TextShape {
-            pos: text_pos,
-            galley: galley.galley,
-            underline: Stroke::NONE,
-            override_text_color,
-            angle: 0.0,
-        });
-
-        let close_response = if show_close {
-            let mut close_button_rect = rect;
-            close_button_rect.set_left(text_rect.right());
-            close_button_rect =
-                Rect::from_center_size(close_button_rect.center(), Vec2::splat(close_button_size));
-
-            let response = ui
-                .interact(close_button_rect, id, Sense::click())
-                .on_hover_cursor(CursorIcon::PointingHand);
-
-            let color = if response.hovered() {
-                self.buttons.close_tab_active_color
-            } else {
-                self.buttons.close_tab_color
-            };
-
-            if response.hovered() {
-                let mut rounding = rounding;
-                rounding.nw = 0.0;
-                rounding.sw = 0.0;
-
-                ui.painter()
-                    .rect_filled(close_button_rect, rounding, self.buttons.add_tab_bg_fill);
-            }
-
-            let mut x_rect = close_button_rect;
-            rect_set_size_centered(&mut x_rect, Vec2::splat(Self::TAB_CLOSE_X_SIZE));
-            ui.painter().line_segment(
-                [x_rect.left_top(), x_rect.right_bottom()],
-                Stroke::new(1.0, color),
-            );
-            ui.painter().line_segment(
-                [x_rect.right_top(), x_rect.left_bottom()],
-                Stroke::new(1.0, color),
-            );
-
-            Some(response)
-        } else {
-            None
-        };
-
-        (response, close_response)
     }
 }
